requirements.txt
aniso8601==8.1.0
astroid==2.4.2
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.8
gunicorn==20.0.4
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pylint==2.6.0
pymongo==3.11.2
python-dotenv==0.15.0
pytz==2020.5
six==1.15.0
toml==0.10.2
Werkzeug==1.0.1
wrapt==1.12.1
